+++ bindings/ruby/ext/css_inline/src/lib.rs
@@ -176,8 +176,25 @@ impl CSSInliner {
     }
 
     #[allow(clippy::needless_pass_by_value)]
+    fn inline_fragment(&self, html: String, css: String) -> RubyResult<String> {
+        Ok(self
+            .inner
+            .inline_fragment(&html, &css)
+            .map_err(InlineErrorWrapper)?)
+    }
+
+    #[allow(clippy::needless_pass_by_value)]
     fn inline_many(&self, html: Vec<String>) -> RubyResult<Vec<String>> {
         inline_many_impl(&html, &self.inner)
+    }
+
+    #[allow(clippy::needless_pass_by_value)]
+    fn inline_many_fragments(
+        &self,
+        html: Vec<String>,
+        css: Vec<String>,
+    ) -> RubyResult<Vec<String>> {
+        inline_many_fragments_impl(&html, &css, &self.inner)
     }
 }
 
@@ -187,6 +204,16 @@ fn inline(args: &[Value]) -> RubyResult<String> {
     let html = args.required.0;
     Ok(rust_inline::CSSInliner::new(options)
         .inline(&html)
+        .map_err(InlineErrorWrapper)?)
+}
+
+fn inline_fragment(args: &[Value]) -> RubyResult<String> {
+    let args = scan_args::<(String, String), _, _, _, _, _>(args)?;
+    let options = parse_options(&args)?;
+    let html = args.required.0;
+    let css = args.required.1;
+    Ok(rust_inline::CSSInliner::new(options)
+        .inline_fragment(&html, &css)
         .map_err(InlineErrorWrapper)?)
 }
 
@@ -205,17 +232,47 @@ fn inline_many_impl(
     Ok(output.map_err(InlineErrorWrapper)?)
 }
 
+fn inline_many_fragments(args: &[Value]) -> RubyResult<Vec<String>> {
+    let args = scan_args::<(Vec<String>, Vec<String>), _, _, _, _, _>(args)?;
+    let options = parse_options(&args)?;
+    let inliner = rust_inline::CSSInliner::new(options);
+    inline_many_fragments_impl(&args.required.0, &args.required.1, &inliner)
+}
+
+fn inline_many_fragments_impl(
+    htmls: &[String],
+    css: &[String],
+    inliner: &rust_inline::CSSInliner<'static>,
+) -> RubyResult<Vec<String>> {
+    let output: Result<Vec<_>, _> = htmls
+        .par_iter()
+        .zip(css)
+        .map(|(html, css)| inliner.inline_fragment(html, css))
+        .collect();
+    Ok(output.map_err(InlineErrorWrapper)?)
+}
+
 #[magnus::init(name = "css_inline")]
 fn init() -> RubyResult<()> {
     let module = define_module("CSSInline")?;
 
     module.define_module_function("inline", function!(inline, -1))?;
+    module.define_module_function("inline_fragment", function!(inline_fragment, -1))?;
     module.define_module_function("inline_many", function!(inline_many, -1))?;
+    module.define_module_function(
+        "inline_many_fragments",
+        function!(inline_many_fragments, -1),
+    )?;
 
     let class = module.define_class("CSSInliner", class::object())?;
     class.define_singleton_method("new", function!(CSSInliner::new, -1))?;
     class.define_method("inline", method!(CSSInliner::inline, 1))?;
+    class.define_method("inline_fragment", method!(CSSInliner::inline_fragment, 2))?;
     class.define_method("inline_many", method!(CSSInliner::inline_many, 1))?;
+    class.define_method(
+        "inline_many_fragments",
+        method!(CSSInliner::inline_many_fragments, 2),
+    )?;
 
     let class = module.define_class("StylesheetCache", class::object())?;
     class.define_singleton_method("new", function!(StylesheetCache::new, -1))?;