+++ bindings/javascript/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - External stylesheet caching. [#314](https://github.com/Stranger6667/css-inline/issues/314)
+- Inlining to HTML fragments. [#335](https://github.com/Stranger6667/css-inline/issues/335)
 
 ## [0.13.2] - 2024-03-25